move generic function into Common, expect localized view in test
